+++ src/convolution/mod.rs
@@ -22,6 +22,7 @@ cfg_if::cfg_if! {
         mod i32x1;
         mod f32x1;
         mod f32x2;
+        mod f32x3;
         mod vertical_u16;
         mod vertical_f32;
     }