+++ src/resizer.rs
@@ -178,6 +178,7 @@ impl Resizer {
             (PT::I32, pixels::I32),
             (PT::F32, pixels::F32),
             (PT::F32x2, pixels::F32x2),
+            (PT::F32x3, pixels::F32x3),
         );
 
         #[cfg(feature = "only_u8x4")]